add comments api module
